Create new dao for retrieving a framework and its days
